revise: classes, inheritence & cmake
I honestly have no idea what is going on with CMake. I thought I
understood it but I suppose not i guess.
* Tried revising Classes & inheritence but CMake had other plans
* Worked more with CMake achieving nothing the whole day :(
* Need to revise my CMake especially aspects of linking libraries &
  linking interface and implementation files.

BREAKING CHANGES:
Everything else works but compiling code with instance of Student
leads to linkage error, interface cannot find implemention of Student
class. This is because these 2 files are in different directories and
my CMake is not configured to link the 2.